Refactor and optimize doctrine status handling
Consolidated redundant functionality and introduced `shared_utils` for shared operations like retrieving and updating doctrine status. Improved modularity and reusability by moving functions such as `fill_missing_stats_v2` and `get_doctrine_status_optimized` to a dedicated module. Enhanced readability and maintainability across the codebase.
